Use System.Reflection.Metadata.AssemblyFlags enum

+++ src/Compilers/Core/Portable/PEWriter/MetadataWriter.cs
@@ -4122,7 +4122,7 @@ namespace Microsoft.Cci
             writer.WriteUInt16((ushort)assembly.Version.Minor);
             writer.WriteUInt16((ushort)assembly.Version.Build);
             writer.WriteUInt16((ushort)assembly.Version.Revision);
-            writer.WriteUInt32(assembly.Flags);
+            writer.WriteUInt32((uint)assembly.Flags);
             writer.WriteReference((uint)heaps.ResolveBlobIndex(_assemblyKey), metadataSizes.BlobIndexSize);
 
             writer.WriteReference((uint)heaps.ResolveStringIndex(_assemblyName), metadataSizes.StringIndexSize);
